+++ infra/infra.yaml
@@ -36,7 +36,15 @@ Parameters:
   S3BucketCodePipelineName:
     Type: String 
     Description: S3 bucket name to store the Bitbucket repository content
-    AllowedPattern: '^[a-z]*$'
+    AllowedPattern: '^((?!xn--)(?!.*-s3alias$)[a-z0-9][a-z0-9-]{1,61}[a-z0-9])$'
+    ConstraintDescription: This field should contain only lower case characters
+  S3BucketRegion:
+    Type: String
+    Description: S3 bucket region  
+  LambdaS3Bucket:
+    Type: String
+    Description: S3 bucket name which stores code for Lambda function
+    AllowedPattern: ^((?!xn--)(?!.*-s3alias$)[a-z0-9][a-z0-9-]{1,61}[a-z0-9])$
     ConstraintDescription: This field should contain only lower case characters
   VPCID:
     Type: AWS::EC2::VPC::Id
@@ -154,12 +162,14 @@ Resources:
             !Ref WebProxyHost
           WEBPROXY_PORT:
             !Ref WebProxyPort
+          S3_BUCKET_REGION:
+            !Ref: S3BucketRegion  
       Handler: index.handler
       KmsKeyArn: !GetAtt KMSKey.Arn
       Role:
         !GetAtt IamRoleLambdaFunction.Arn
       Code: ../code
-      Runtime: nodejs14.x
+      Runtime: nodejs20.x
       Timeout: '30'
       VpcConfig:
         SecurityGroupIds:
